PER-53 #time changed datatype of hsv color

+++ suturo_perception_lib/include/perceived_object.h
@@ -17,11 +17,11 @@ namespace suturo_perception_lib
       uint8_t c_color_average_r;
       uint8_t c_color_average_g;
       uint8_t c_color_average_b;
-      uint8_t c_color_average_h;
-      uint8_t c_color_average_s;
-      uint8_t c_color_average_v;
+      uint32_t c_color_average_h;
+      double c_color_average_s;
+      double c_color_average_v;
       std::string c_recognition_label_2d;
-      std::vector<int> c_hue_histogram;
+      std::vector<uint32_t> c_hue_histogram;
       uint8_t c_hue_histogram_quality;
       ROI c_roi;
       pcl::PointCloud<pcl::PointXYZRGB>::Ptr pointCloud;
